feat: add model classes Filme, Serie, and Titulo in modelos package
