+++ src/br/com/alura/screenmatch/modelos/Serie.java
@@ -1,0 +1,46 @@
+package br.com.alura.screenmatch.modelos;
+
+public class Serie extends Titulo {
+
+    private int temporadas;
+    private boolean ativa;
+    private int episodiosPorTemporada;
+    private  int minutosPorEpisodios;
+
+    public int getTemporadas() {
+        return temporadas;
+    }
+
+    public void setTemporadas(int temporadas) {
+        this.temporadas = temporadas;
+    }
+
+    public boolean isAtiva() {
+        return ativa;
+    }
+
+    public void setAtiva(boolean ativa) {
+        this.ativa = ativa;
+    }
+
+    public int getEpisodiosPorTemporada() {
+        return episodiosPorTemporada;
+    }
+
+    public void setEpisodiosPorTemporada(int episodiosPorTemporada) {
+        this.episodiosPorTemporada = episodiosPorTemporada;
+    }
+
+    public int getMinutosPorEpisodios() {
+        return minutosPorEpisodios;
+    }
+
+    public void setMinutosPorEpisodios(int minutosPorEpisodios) {
+        this.minutosPorEpisodios = minutosPorEpisodios;
+    }
+
+    @Override
+    public int getDuracaoEmMinutos() {
+        return temporadas * episodiosPorTemporada * minutosPorEpisodios;
+    }
+}
